RainbowCORE 1.1
Resolved the following issues...

Styles no longer working after Firefox upgrade
Cause:
After investigating, this seemed to be because the style script was running too quickly, and was finish running before it could effect the elements being loaded in.

Resolution
Added a timeout of 10 milliseconds to RainbowCORE_style.user.js

After using column searching, first column becomes incorrect color
Cause:
This is actually an issue with Reliance where it adds the "SelectedViewRecord" class to the first record after a search (even if the user has not actually selected that record).

It also seems that there is are insistences when the class is added by clicking on the search bar. Looking at the main Reliance.JS file within Reliance, this does not look intentional and is an issue with OOB Reliance.

Resolution:
I added a section in RainbowCORE_main.user.js to remove the "SelectedViewRecord" class when the document originally loads. However, there will not be a fix through RainbowCORE to fix the intermittent bugs of the class being added when clicking the search bar. This would be require too many events and take up too many resources to resolve client side.

Attached images no longer working. Images show as blank instead of displaying
Cause:
The RainbowCORE_style.user.js script running before the related attachment image caused the attachment image to no longer load. It seemed to be that any script running on the document start broke the TLS handshake between the image and reliance. It had nothing to do with the script itself, but just that the script was running because other resources was loaded in

Resolution:
Added an exclude to the files to not run on attachment windows since they don't need to run on those anyway.